add tableview/delegate/datasource/structure of project
