+++ STARTIOS-HW13-Zhantassov Adlet/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
