+++ STARTIOS-HW13-Zhantassov Adlet/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
