+++ STARTIOS-HW13-Zhantassov Adlet/Resources/Assets.xcassets/Contents.json
