+++ STARTIOS-HW13-Zhantassov Adlet/Resources/Configuration/Info.plist
