+++ STARTIOS-HW13-Zhantassov Adlet/Sources/App/AppDelegate.swift
