+++ STARTIOS-HW13-Zhantassov Adlet/Sources/App/SceneDelegate.swift
