+++ STARTIOS-HW13-Zhantassov Adlet/Sources/Storyboards/Base.lproj/LaunchScreen.storyboard
