+++ STARTIOS-HW13-Zhantassov Adlet/Sources/UIKit+Extensions/Ext+TableView.swift
@@ -1,0 +1,20 @@
+//
+//  File.swift
+//  STARTIOS-HW13-Zhantassov Adlet
+//
+//  Created by Adlet Zhantassov on 05.03.2023.
+//
+
+import Foundation
+import UIKit
+
+extension ViewController: UITableViewDelegate, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        10
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
+        return cell
+    }
+}
