--- STARTIOS-HW13-Zhantassov Adlet/ViewController.swift
@@ -1,35 +1,0 @@
-//
-//  ViewController.swift
-//  STARTIOS-HW13-Zhantassov Adlet
-//
-//  Created by Adlet Zhantassov on 05.03.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-    
-    //MARK: - SetupData
-    
-    
-    
-    //MARK: - Lifecycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupViews()
-        setupConstraints()
-    }
-    
-    //MARK: - SetupViews
-    private func setupViews() {
-        view.backgroundColor = .white
-    }
-    
-    //MARK: - SetupConstraints
-    private func setupConstraints() {
-        
-    }
-
-
-}
-
